pyproject.toml
[project]
name = "pycinemetrics"
version = "0.1.0"
description = "Scene-centric film analysis toolkit with V2 web UI and backend API."
readme = "readme.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "easyocr>=1.7.2",
    "ffmpeg-python>=0.2.0",
    "fastapi>=0.116.1",
    "jieba>=0.42.1",
    "matplotlib>=3.10.5",
    "numpy>=2.1.3",
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
    "tensorflow>=2.19.0",
    "torchvision>=0.23.0",
    "transnetv2",
    "uvicorn>=0.35.0",
    "wordcloud>=1.9.4",
    "python-multipart>=0.0.20",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "ruff>=0.13.1",
]

[tool.uv.sources]
transnetv2 = { git = "https://github.com/soCzech/TransNetV2" }

[tool.pytest.ini_options]
pythonpath = ["app"]
testpaths = []

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = [".venv", "models", "img"]

[tool.ruff.lint]
select = ["F"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
mypy_path = "app"
exclude = "(^models/|^img/)"
